requirements.txt
anyio==4.12.1
asgiref==3.11.1
attrs==25.4.0
bcrypt==5.0.0
beautifulsoup4==4.14.3
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.5
dj-database-url==3.1.2
Django==6.0.3
cloudinary==1.44.1
django-cloudinary-storage==0.3.0
django-cors-headers==4.9.0
django-filter==24.3
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
duckduckgo_search==8.1.1
google-auth==2.48.0
google-auth-oauthlib==1.3.0
googlesearch-python==1.3.0
gspread==6.2.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
inflection==0.5.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
lxml==6.0.2
nest-asyncio==1.6.0
oauthlib==3.3.1
primp==1.1.2
psycopg2-binary==2.9.11
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
PyJWT==2.11.0
python-dotenv==1.2.2
python-telegram-bot==22.6
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
requests-oauthlib==2.0.0
rpds-py==0.30.0
rsa==4.9.1
soupsieve==2.8.3
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.3
uv==0.10.9
whitenoise==6.12.0
gunicorn==25.1.0
